feat(dj): Add library context menu and switch to SoundTouch BPM detection
- Add right-click context menu to library tracks with:
  - Re-analyze BPM (queues track for background analysis)
  - Edit BPM manually (opens dialog to enter custom value)
  - Delete from library
  - Show in Finder (macOS) / Show in File Manager (Linux)
- Switch BPM detection from rustfft autocorrelation to SoundTouch BPMDetect
- Fix library not updating BPM values after track analysis completes
- Add file_path to DjTrackInfo for "Show in Finder" feature

+++ crates/core/src/modules/traits.rs
@@ -113,7 +113,7 @@ pub enum ModuleEvent {
         loop_in: Option<f64>,
         loop_out: Option<f64>,
         active: bool,
-        beat_count: u8,
+        beat_count: f64,
     },
     /// DJ track analysis progress (background import)
     DjAnalysisProgress {
